Split out the binding classes specific to CMP from the controller to reduce the file size.
