+++ paf-mvp-cmp/src/controller.ts
@@ -1,5 +1,6 @@
 import { Locale } from './locale';
 import { Config } from './config';
+import { BindingShowRandomId, BindingThisSiteOnly } from './bindings';
 import { Log } from '@core/log';
 import { BindingButton, BindingChecked, BindingCheckedMap, BindingElement, BindingViewOnly } from '@core/ui/binding';
 import { Identifier, IdsAndOptionalPreferences, Preferences, PreferencesData } from '@core/model/generated-model';
@@ -556,60 +557,6 @@ class BindingDisplayRandomId extends BindingViewOnly<Identifier, Model, HTMLSpan
   public bind(): void {
     if (this.field !== null) {
       this.setValue(this.field.value);
-    }
-  }
-}
-
-/**
- * Hides the this site only option if the feature is not configured.
- */
-class BindingThisSiteOnly extends BindingViewOnly<boolean, Model, HTMLDivElement> {
-  private readonly enabled: boolean;
-
-  constructor(view: View, id: string, config: Config) {
-    super(view, id);
-    this.enabled = config.siteOnlyEnabled;
-  }
-
-  public bind(): void {
-    const element = this.getElement();
-    if (element !== null) {
-      element.style.display = this.enabled ? '' : 'none';
-    }
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  setValue(value: boolean): void {
-    // Do nothing.
-  }
-}
-
-/**
- * Custom UI binding to hide or show the area that displays the random identifier if preferences have been set.
- */
-class BindingShowRandomId extends BindingViewOnly<PreferencesData, Model, HTMLDivElement> {
-  protected readonly model: Model;
-
-  constructor(view: View, id: string, model: Model) {
-    super(view, id);
-    this.model = model;
-  }
-
-  public bind(): void {
-    if (this.field !== null) {
-      this.setValue(this.field.value);
-    }
-  }
-
-  /**
-   * If the preferences are persisted then show the identifier.
-   * @param value of the identifier being displayed
-   */
-  public setValue(value: PreferencesData) {
-    const element = super.getElement();
-    if (element !== null) {
-      const visible = value !== null && this.model.rid?.value?.value !== undefined;
-      element.style.display = visible ? '' : 'none';
     }
   }
 }
